% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_named_listener}
\alias{add_named_listener}
\title{Add named listener to an event}
\usage{
add_named_listener(event, name, listener)
}
\arguments{
\item{event}{a [individual::Event] or [individual::TargetedEvent] object}

\item{name}{a name}

\item{listener}{a function which takes either a single argument (`timestep`)
if event is a [individual::Event] or two arguments (`timestep`, `target`) if
the event is a [individual::TargetedEvent]}
}
\description{
Given a [individual::Event] or [individual::TargetedEvent] object,
add a named listener to the list of listeners stored in that event.
}
