% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_compartments.R
\name{create_compartment_variables}
\alias{create_compartment_variables}
\title{Create compartmental state variables}
\usage{
create_compartment_variables(psq)
}
\arguments{
\item{psq}{the model parameters}
}
\value{
list of states
}
\description{
Create_states creates and initialises the human states for the model
}
