% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_base.R
\name{create_events}
\alias{create_events}
\title{Create events}
\usage{
create_events(parameters)
}
\arguments{
\item{parameters}{model parameters}
}
\description{
Create named list of \link[individual:TargetedEvent]{individual::TargetedEvent} objects which
simulate disease progression. The list of events is as follows:
\itemize{
\item \code{exposure}: scheduled when an individual is infected
\item \code{mild_infection}: scheduled from \code{exposure}
\item \code{asymp_infection}: scheduled from \code{exposure}
\item \code{severe_infection}: scheduled from \code{exposure}
\item \code{hospitilisation}: scheduled from \code{severe_infection}
\item \code{ICU_get_live}: scheduled from \code{hospitilisation}
\item \code{ICU_get_die}: scheduled from \code{hospitilisation}
\item \code{hosp_get_live}: scheduled from \code{hospitilisation}
\item \code{hosp_get_die}: scheduled from \code{hospitilisation}
\item \code{ICU_not_get_live}: scheduled from \code{hospitilisation}
\item \code{ICU_not_get_die}: scheduled from \code{hospitilisation}
\item \code{hosp_not_get_live}: scheduled from \code{hospitilisation}
\item \code{hosp_not_get_die}: scheduled from \code{hospitilisation}
\item \code{stepdown}: scheduled from \code{ICU_get_live}
\item \code{recovery}: scheduled from \code{mild_infection}, \code{asymp_infection}, \code{ICU_not_get_live}, \code{hosp_get_live}, \code{hosp_not_get_live}
\item \code{immunity_loss}: scheduled from \code{recovery}
\item \code{death}: scheduled from \code{ICU_get_die}, \code{ICU_not_get_die}, \code{hosp_get_die}, \code{hosp_not_get_die}
}
}
