% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_events.R
\name{create_state_update_listener}
\alias{create_state_update_listener}
\title{A listener to update state}
\usage{
create_state_update_listener(states, destination)
}
\arguments{
\item{states}{a \link[individual:CategoricalVariable]{individual::CategoricalVariable} object}

\item{destination}{the destination state}
}
\description{
a listener to be attached to a \link[individual:TargetedEvent]{individual::TargetedEvent}
to update state when that event fires.
}
