% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{interpolate_rt}
\alias{interpolate_rt}
\title{Produce piecewise linear Rt}
\usage{
interpolate_rt(dates, rt, max_date = NULL)
}
\arguments{
\item{dates}{a vector of [Date]s}

\item{rt}{a vector of reproductive values at those dates}

\item{max_date}{the maximum date of simulation (optional)}
}
\value{
a [list] with named elements `Rt` and `Rt_tt`, which can be provided
to the [safir::get_parameters] function's arguments `R0` and `tt_R0`.
}
\description{
Make a piecewise linear interpolating vector of Rt, for each
day within the simulated interval. The assumed start of simulation
is the first date provided in `dates`, and the end of simulation is the last
date of `dates` unless `max_date` is provided. Interpolation is done with
inclusive endpoints.
}
