% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_get_age_probabilities}
\alias{make_get_age_probabilities}
\title{Get age-structured transition probabilities at current timestep}
\usage{
make_get_age_probabilities(parameters, name)
}
\arguments{
\item{parameters}{a named [list]}

\item{name}{an element in the parameter list, must be numeric}
}
\description{
Return a function taking two arguments `timestep` and `ages` which
returns the probabilities stored in `name` at the current timestep and for
those persons who age is in `ages`.
}
