% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parameters.R
\name{make_get_contact_matrix}
\alias{make_get_contact_matrix}
\title{Get contact matrix at current timestep}
\usage{
make_get_contact_matrix(parameters)
}
\arguments{
\item{parameters}{a named \link{list}}
}
\description{
Return a function taking a single argument \code{timestep} which
returns a contact matrix for the current timestep
}
