% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_calculations.R
\name{make_get_NAT}
\alias{make_get_NAT}
\title{Make function to calculate population NAT}
\usage{
make_get_NAT(variables)
}
\arguments{
\item{variables}{a named list}
}
\value{
a function
}
\description{
This returns a function taking a single argument
\code{target} (a \link[individual:Bitset]{individual::Bitset}) which will return the natural log scale NAT for each person in \code{index}.
}
