% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_calculations.R
\name{make_get_ve_severe}
\alias{make_get_ve_severe}
\title{Make function to calculate efficacy against severe disease}
\usage{
make_get_ve_severe(variables, parameters)
}
\arguments{
\item{variables}{a named list}

\item{parameters}{a named list}
}
\value{
a function
}
\description{
Return a function taking two arguments \code{timestep} and \code{target} (a \link[individual:Bitset]{individual::Bitset})
which will calculate the efficacy against severe disease from neutralizing antibody titre.
}
