% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parameters.R
\name{make_get_vector}
\alias{make_get_vector}
\title{Get vector or scalar parameter at current timestep}
\usage{
make_get_vector(parameters, name)
}
\arguments{
\item{parameters}{a named \link{list}}

\item{name}{an element in the parameter list, must be numeric}
}
\description{
Return a function taking a single argument \code{timestep} which
returns the value of \code{name} at the current timestep
}
