% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safir2.R
\docType{package}
\name{safir2}
\alias{safir2}
\alias{safir2-package}
\title{safir2: squire and friends individual rewrite rewrite}
\description{
'safir2' is an upgraded and refactored version of 'safir', which implements a stochastic individual
 immunity and neutralizing antibody titre, as well as allowing for complex vaccine
 allocation schemes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/safir2}
  \item \url{https://mrc-ide.github.io/safir2/}
  \item Report bugs at \url{https://github.com/mrc-ide/safir2/issues}
}

}
\author{
\strong{Maintainer}: Sean L. Wu \email{slwood89@gmail.com} (\href{https://orcid.org/0000-0002-5781-9493}{ORCID})

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
