% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_calculations.R
\name{ve_infection_functional_form}
\alias{ve_infection_functional_form}
\title{Function for efficacy against infection}
\usage{
ve_infection_functional_form(NAT, k, ab_50)
}
\arguments{
\item{NAT}{neutralizing antibody titre on a linear scale}

\item{k}{half maximum parameter}

\item{ab_50}{scaling parameter}
}
\description{
\deqn{1 - \left( \frac{1}{1 + e^{-k \cdot \log_{10}(NAT) - \log_{10}(ab_{50})}} \right) }
}
