% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{add_named_listener}
\alias{add_named_listener}
\title{Add named listener to an event}
\usage{
add_named_listener(event, name, listener)
}
\arguments{
\item{event}{a \link[individual:Event]{individual::Event} or \link[individual:TargetedEvent]{individual::TargetedEvent} object}

\item{name}{a name}

\item{listener}{a function which takes either a single argument (\code{timestep})
if event is a \link[individual:Event]{individual::Event} or two arguments (\code{timestep}, \code{target}) if
the event is a \link[individual:TargetedEvent]{individual::TargetedEvent}}
}
\description{
Given a \link[individual:Event]{individual::Event} or \link[individual:TargetedEvent]{individual::TargetedEvent} object,
add a named listener to the list of listeners stored in that event.
}
